requirements.txt
httpx>=0.27.2
python-dotenv>=1.0.1
rich>=13.9.4
caldav==1.4.0
peewee==3.17.8
schedule==1.2.2